EditComment and optional automatic EditComment call
Added support for EditComment as well as a second override to
NewComment. In most WordPress installations, anonymous comments over
XML-RPC are not allowed. This causes all new comments to be posted as
the username used to call NewComment. to fix this, either create a
WordPress plugin to allow anonymous comments (very bad), call the
NewComment with the optional boolean parameter at the end, or call
EditComment manually after calling NewComment.